chore(architecture): move files for new architecture
- Move ensemble class for new project structure (mlabican lib);

Issues: #16, #22 and #27
